adding folder structure
