+++ src/Pages/cs.css
@@ -1,0 +1,46 @@
+.bg{
+    background: url(https://www.w3schools.com/w3images/forestbridge.jpg);
+    min-height: 100%;
+    background-position: center;
+    background-size: cover;
+}
+
+#root{
+    height: 100%;
+}
+
+.logo {
+    padding: 12px 24px!important;
+    font-size: 24px!important;
+    color: white;
+    font-family: "Raleway", sans-serif;
+}
+
+.center{
+    color: white;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%,-50%);
+}
+
+.cs{
+    font-size: 64px !important;
+    position: relative;
+    animation: animatetop 0.4s;
+    font-weight: 400 !important;
+    margin: 10px 0 !important;
+}
+
+.ms{
+    text-align: center!important;
+    font-size: 18px!important;
+}
+
+.footer{
+    padding: 12px 24px!important;
+    color: white;
+    position: absolute;
+    left: 0;
+    bottom: 0;
+}
