+++ src/Pages/cs.js
@@ -1,0 +1,25 @@
+import React from 'react'
+import './cs.css'
+
+const cs = () => {
+    return (
+        <>
+            
+<div className="bg">
+  <div className="logo">
+    Logo
+  </div>
+  <div className="center">
+    <h1 className="cs">COMING SOON</h1>
+    <hr className=""/>
+    <p className="ms">35 days left</p>
+  </div>
+  <div className="footer">
+    Powered by Deepak & Shubham
+  </div>
+</div>
+        </>
+    )
+}
+
+export default cs;
